requirements.txt
python-docx
BeautifulSoup
selenium
streamlit
